requirements.txt
requests
google-generativeai
gtts
moviepy
google-api-python-client
google-auth-httplib2
google-auth-oauthlib